fse: rename tips folder
