+++ apps/full-site-editing/full-site-editing-plugin/block-inserter-modifications/contextual-tips/style.scss
