+++ apps/full-site-editing/full-site-editing-plugin/block-inserter-modifications/index.php
@@ -74,8 +74,8 @@ function enqueue_block_inserter_modifications() {
 	 */
 	enqueue_script( 'new-blocks-showcase', false );
 
-	enqueue_script( 'tips', true );
-	enqueue_style( 'tips', false );
+	enqueue_script( 'contextual-tips', true );
+	enqueue_style( 'contextual-tips', false );
 }
 
 add_action( 'enqueue_block_editor_assets', __NAMESPACE__ . '\enqueue_block_inserter_modifications', 0 );
